requirements.txt
langchain>=0.1.16
langchain-community>=0.0.34
langchain-openai>=0.1.3
openai>=1.23.2
chromadb>=0.4.24
pypdf>=4.2.0
tiktoken>=0.6.0
gradio>=4.26.0
python-dotenv>=1.0.1
requests>=2.32.4
opentelemetry-api>=1.37.0
opentelemetry-sdk>=1.37.0
opentelemetry-exporter-otlp-proto-common>=1.37.0
opentelemetry-proto>=1.37.0
